+++ lua/terrortown/lang/en.lua
@@ -2131,7 +2131,7 @@ L.help_roles_builtin = "This role is built-in, it comes with TTT2!"
 L.header_equipment_info = "Equipment information"
 
 
--- 2023-12-22
+-- 2023-12-24
 L.submenu_gameplay_accessibility_title = "Accessibility"
 
 L.header_accessibility_settings = "Accessibility Settings"
@@ -2142,7 +2142,6 @@ L.label_enable_bobbing_strafe = "Enable view bobbing when strafing"
 
 L.help_enable_dynamic_fov = "Dynamic FOV is applied depending on the player's speed. When a player is sprinting for example, the FOV is increased to visualize the speed."
 L.help_enable_bobbing_strafe = "View bobbing is the slight camera shake while walking, swimming or falling."
-
 -- 2023-12-20
 L.equipmenteditor_desc_damage_scaling = [[Multiplies the base damage value of a weapon by this factor.
 For a shotgun, this would affect each pellet.
@@ -2156,6 +2155,8 @@ Note: Some weapons might not use this value which causes this modifier to be ine
 
 -- 2023-12-24
 L.binoc_help_reload = "Clear target."
+L.cl_sb_row_sresult_direct_conf = "Direct confirmation"
+L.cl_sb_row_sresult_pub_police = "Public policing role confirmation"
 
 -- 2024-01-05
 L.label_crosshair_thickness_outline_enable = "Enable crosshair outline"